Add files via upload
Contains all packages required to run the python code for the analyses run in the Jupyter Notebooks. Please run Python pip in command line using the following line: "pip install -r requirements.txt" to use this file correctly (as this line with this requirements file) will install the packages with specific versions that work with this code to the local machine so that the analyses can be run on a local machine.